Different display and 8 encoders.
1 encoder doesn't properly work though.
Need to investigate.
Not only the button pin interferes with Serial In
(which is unused really but still messes us up somehow -
but anyway, Serial debugging can be disabled in prod),
but one of the pins is always getting a 0 for some reason.
I thought it was a defective pin but I checked on a different
Blue Pill board and it's the same there.