Added bottom part of the enclosure.
The contents barely fit and the bottom part visibly bends. I'm lucky I
was able to 3D-print with little warping, so I was able to successfully
use metal screws on the side. But in fugure I'd try thicker bottom
floor, taller lip for better friction, or have more screw holes.