Added Ableton Live scripts.
Ableton Live integration is based on Python scripts that Ableton uses for control surfaces.

I based mine on Novation SL MkIII since it also has an LCD with 8 parameters mapped to encoders.

MIDI sysex messages are used to update the display. For the most part I'm using the same MIDI messages as SL MkIII.

This can probably be cleaned up and refactored but absolutely despise reverse-engineering Ableton's Python scripts
and it's already working, so for now I'll just commit it as it is.